requirements.txt
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
httpx>=0.23.0,<0.25.0
google-generativeai==0.3.1
Pillow==10.1.0
GitPython==3.1.40
markdown==3.5.1
pyyaml==6.0.1
python-frontmatter==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3 